fix: Update file upload middleware to create directories recursively

+++ src/utils/cleanFile.ts
@@ -7,6 +7,13 @@ export default async function cleanFile(fileName: string) {
 			return log.warn(
 				"A file was not cleaned because it was not provided"
 			)
+
+		if (!fs.existsSync(fileName))
+			return log.warn(
+				"A file was not cleaned because it does not exist: ",
+				fileName
+			)
+
 		await fs.promises.unlink(fileName)
 		log.info("A file was cleaned: ", fileName)
 	} catch (err) {
